requirements.txt
moviepy
pydub
openai-whisper
python-dotenv
numpy
